[REFACTOR] 实装重构的ResourceManager. （旧的被我用namespace包裹了），解决因为实装而导致的报错。

+++ Assets/Scripts/Tools/FunctionTester.cs
@@ -25,12 +25,12 @@ public class FunctionTester : MonoBehaviour
     public void TestResourceAddRemove()
     {
         Debug.Log("[FunctionTester] ===============================测试资源添加和移除===============================");
-        ResourceManager.Instance.AddResource(ResourceType.Seed, SeedSubType.Wheat, 10); // 正常添加
-        ResourceManager.Instance.RemoveResource(ResourceType.Seed, SeedSubType.Wheat, 10); // 正常移除
-        ResourceManager.Instance.AddResource(ResourceType.Seed, SeedSubType.Wheat, 9999); // 添加超过存储上限的资源
-        ResourceManager.Instance.RemoveResource(ResourceType.Seed, SeedSubType.Corn, 10); // 移除不存在的资源
-        ResourceManager.Instance.AddResource(ResourceType.Seed, SeedSubType.Corn, 10); // 添加不存在的资源
-        ResourceManager.Instance.RemoveResource(ResourceType.Seed, SeedSubType.Wheat, 9999); // 移除超过资源数量的资源
+        ResourceManager.Instance.AddResource(ResourceType.Seed, (int)SeedSubType.Wheat, 10); // 正常添加
+        ResourceManager.Instance.RemoveResource(ResourceType.Seed, (int)SeedSubType.Wheat, 10); // 正常移除
+        ResourceManager.Instance.AddResource(ResourceType.Seed, (int)SeedSubType.Wheat, 9999); // 添加超过存储上限的资源
+        ResourceManager.Instance.RemoveResource(ResourceType.Seed, (int)SeedSubType.Corn, 10); // 移除不存在的资源
+        ResourceManager.Instance.AddResource(ResourceType.Seed, (int)SeedSubType.Corn, 10); // 添加不存在的资源
+        ResourceManager.Instance.RemoveResource(ResourceType.Seed, (int)SeedSubType.Wheat, 9999); // 移除超过资源数量的资源
     }
 
     [ContextMenu("Test Hire NPC")]
